+++ app/src/main/java/com/billdoerr/android/carputer/activities/MainActivity.java
@@ -1,8 +1,11 @@
 package com.billdoerr.android.carputer.activities;
 
+import android.os.Bundle;
+
 import androidx.fragment.app.Fragment;
 
 import com.billdoerr.android.carputer.fragments.CameraFragmentMotionEye;
+import com.billdoerr.android.carputer.utils.GlobalVariables;
 
 /**
  * Carputer activity which extends from BaseActivity.
@@ -12,6 +15,17 @@ public class MainActivity extends BaseActivity {
     @Override
     protected Fragment createFragment() {
         return CameraFragmentMotionEye.newInstance();
+    }
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        GlobalVariables globalVariables = (GlobalVariables) getApplicationContext();
+        boolean enabled = globalVariables.isBottomNavigationEnabled();
+
+        // Display BottomNavigationView
+        setBottomNavigationViewVisibility(enabled);
     }
 
 }